完成一部分patch
